Remove unnecessary abstraction layer `ValueSpec`
The functionality was used only by the `Port` class and it simply
constructed an instance in its constructor and forwarded the property
getter and setters. This unnecessary layer is removed and the
functionality is directly implemented on the `Port` base class.
